fixes for gcc 2.95.2 and 3.4
git-svn-id: svn://svn.powerdns.com/pdns/trunk/pdns@193 d19b8d6e-7fed-0310-83ef-9ca221ded41b

+++ pdns/packetcache.cc
@@ -197,7 +197,7 @@ map<char,int> PacketCache::getCounts()
     key=i->first[offset+1];
     if((key=='Q' || key=='q') && !i->second.value.empty())
       key='!';
-    counts[key]++;
+    counts[(int)key]++;
   }
 
   map<char,int>ret;
